26/11, added a lot of classes, tried to implement junction detection and linefollowing to no sucess

+++ SearchAndRescueRobot/include/MotorDriver.h
@@ -4,12 +4,12 @@
 class MotorDriver {
     public:
     
-     void begin(int leftA, int leftB, int rightA, int rightB);
+     void begin(uint8_t leftA, uint8_t leftB, uint8_t rightA, uint8_t rightB);
      void drive(int leftSpeed, int rightSpeed);
      void stop();
 
     private:
-     int _lA, _lB, _rA, _rB;
+     uint8_t _lA, _lB, _rA, _rB;
     
 
 };
